+++ src/pages/Faucet/Faucet.tsx
@@ -21,7 +21,7 @@ export const Faucet = () => {
     <>
       <Button
         className='inline-block rounded-lg bg-blue-600 px-4 py-2 text-sm text-white'
-        data-testid={DataTestIdsEnum.sovereignTransferBtn}
+        data-testid={DataTestIdsEnum.faucetBtn}
         onClick={handleShow}
       >
         Request funds